add react-modal as popup window
