+++ william-profile/src/components/Project.js
@@ -1,34 +1,9 @@
-import React from 'react';
+import React, {useState} from 'react';
 import projects from '../assets/datas/ProjectData.json';
 import translations from '../assets/languages/translations.json';
 import { useLanguage } from './LanguageContext';
+import ProjectCard from './ProjectCard';
 
-function ProjectCard(props) {
-
-    return (
-        <div className='projectCard'>
-            <div className='projectTitleTop'>
-                            <div>
-                                <h4 className='projectTitle' >{props.data.title}</h4>
-                            </div>
-                            <div>
-                                <p>{props.data.status}</p>
-                            </div>
-                        </div>
-            <p>{props.data.brief}</p>
-            <h5>Project Details</h5>
-            <ul>
-                {props.data.tasks.map((task) => {
-                    return (
-                        <li key={task.id}>{task.task}</li>
-                    );
-                })}
-            </ul>
-            <p>More Details would be provided with the updating of this site</p>
-        </div>
-
-    );
-}
 
 function Project() {
     const { language } = useLanguage();
@@ -40,14 +15,13 @@ function Project() {
 
                 <div className='projectCardsHolder'>
 
-                    {projects[language].map((project, index) => {
+                    {projects[language].map((project) => {
                         return (
-                            <ProjectCard data={project} key={index} />
-                        );
-                    }
-                    )}
+                            <ProjectCard key={project.id} data={project} />
+                            );
+                        }
+                        )}
                 </div>
-
             </div>
         </section>
     );
